Add TS to greenboard
